+++ python/metatomic_torch/metatomic/torch/model.py
@@ -575,21 +575,8 @@ class AtomisticModel(torch.nn.Module):
         if self._capabilities.length_unit == "":
             warnings.warn(
                 "No length unit was provided for the model.",
-                stacklevel=1,
+                stacklevel=2,
             )
-
-        for name, output in self._capabilities.outputs.items():
-            # TODO: coordinate a list of standard outputs needing
-            # unit checks, should also be consistent with `outputs.py`
-            if name in ["energy", "energy_ensemble", "energy_uncertainty"]:
-                if output.unit == "":
-                    warnings.warn(
-                        f"No units were provided for output {name}.",
-                        stacklevel=1,
-                    )
-
-        # TODO: can we freeze these?
-        # module = torch.jit.freeze(module)
 
         # Metadata about where and when the model was exported
         export_metadata = {