+++ metatomic-torch/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to metatomic-torch are documented here, following the [keep
 a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- Possible sections for each package:
+<!-- Possible sections:
 ### Added
 
 ### Fixed
@@ -23,12 +23,14 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
   `(eV*u)^(1/2)`, etc.) with automatic dimensional validation
 - 2-argument `unit_conversion_factor(from_unit, to_unit)` that parses
   arbitrary unit expressions and checks dimensional compatibility
+- Added `unit_dimension_for_quantity()` to get the physical dimension (i.e.
+  `energy`, `length`, …) corresponding to a standard quantity (output/input) in
+  metatomic.
 
 ### Changed
 
 - 3-argument `unit_conversion_factor(quantity, from_unit, to_unit)` is
   deprecated; the `quantity` parameter is ignored
-
 - `metatomic.torch.ase_calculator` has been split into a separate
   `metatomic-ase` package. The code is temporarily re-exported from the old
   path, but all users are encouraged to update to explicitly requiring